move unrequired functions from main to app to follow SOLID
