+++ rental.rb
@@ -4,11 +4,9 @@ class Rental
 
   def initialize(date, book, person)
     @date = date
-
     @book = book
-    book.rentals << self
-
     @person = person
+    book.rentals << self
     person.rentals << self
   end
 end
